Always use uint8 for metadata and control messages
This greatly simplifies the interfaces specifications, so that
they no longer need to be generic. Everything is just byte arrays.

For encoding/decoding messages, the implementor needs to provide
simple conversion routines from bytes to the desired type. This
shouldn't be a problem for most languages, but might require
unsafe Rust. We'll have to see how it pans out in the long term.