Simplify Multiplexer interface
* renamed sendControlMessage to sendMessage
* renamed onControlMessage to onMessage
* removed setSendHandler
* removed handleReceivedMessage

Initially Multiplexer was designed to be flexible so the owner could handle the transport
layer and pass the messages to be handled to the Mux. However, over time the
implementations have gravitated toward having a concrete Transport interface, an
implemetor of which gets passed in to the Mux. If I recall correctly I originally started
doing it this way to work easier with Rust's ownership. It provides a nice mental
model too, though. You don't need to know the Mux interface in order to implement
a Transport.